--- src/components/Footer.jsx
@@ -1,7 +1,0 @@
-const Footer = () => {
-    return (
-        <p>This is the footer</p>
-    )
-}
-
-export default Footer
