+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,9 @@
+const Footer = () => {
+    return (
+        <footer>
+            <p>This is the footer</p>
+        </footer>
+    )
+}
+
+export default Footer
