--- src/components/Header.jsx
@@ -1,7 +1,0 @@
-const Header = () => {
-    return (
-        <h1>📰 NC News</h1>
-    )
-}
-
-export default Header
