+++ src/components/Header/Header.jsx
@@ -1,0 +1,11 @@
+import styles from "./Header.module.css"
+
+const Header = () => {
+    return (
+        <nav>
+            <h1 className={styles.header}>NC News</h1>
+        </nav>
+    )
+}
+
+export default Header
