+++ src/components/Header/Header.module.css
@@ -1,0 +1,11 @@
+.header {
+    font-size: 3rem;
+    font-family: "Menlo";
+    font-weight: bold;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+    background-clip: text;
+    background-image: url(../../..//public/images/background.jpg);
+    background-position: center bottom;
+    text-align: center;
+}
