+++ src/components/Home/ArticleGrid/ArticleGrid.jsx
@@ -1,6 +1,6 @@
-import { ArticleCard } from "./"
+import ArticleCard from "./ArticleCard"
 import { useState, useEffect } from "react"
-import { fetchArticles } from '../utils/api.js'
+import { fetchArticles } from '../../../utils/api.js'
 
 const ArticleGrid = () => {
     const [articles, setArticles] = useState([])
