+++ src/components/Home/Home.jsx
@@ -1,11 +1,12 @@
-import { Filter, ArticleGrid } from "./"
+import Filter from "./Filter"
+import ArticleGrid from "./ArticleGrid/ArticleGrid"
 
 const Home = () => {
     return (
-        <>
+        <main>
             <Filter />
             <ArticleGrid />
-        </>
+        </main>
     )
 }
