--- src/components/Login.jsx
@@ -1,7 +1,0 @@
-const Login = () => {
-    return (
-        <p>This is the login</p>
-    )
-}
-
-export default Login
