+++ src/components/Login/Login.jsx
@@ -1,0 +1,9 @@
+const Login = () => {
+    return (
+        <main>
+            <p>This is the login</p>
+        </main>
+    )
+}
+
+export default Login
