--- src/components/index.jsx
@@ -1,16 +1,0 @@
-
-// all urls
-export {default as Header} from "./Header"
-export {default as Footer} from "./Footer"
-
-// home page (/)s
-export {default as Home} from "./Home"
-export {default as Filter} from "./Filter"
-export {default as ArticleGrid} from "./ArticleGrid"
-export {default as ArticleCard} from "./ArticleCard"
-
-// topics page (/topics)
-export {default as Topics} from "./Topics"
-
-// login page (/login)
-export {default as Login} from "./Login"
